+++ tools/antithesis-go-instrumentor/testdata/input_multimodule/main.go
@@ -1,0 +1,14 @@
+package main
+
+import (
+	"fmt"
+
+	"example.com/multimodule-test/sub"
+	"github.com/antithesishq/antithesis-sdk-go/assert"
+)
+
+func main() {
+	fmt.Println("hello from root")
+	assert.Always(true, "root always", nil)
+	sub.DoIt()
+}